Add SVG Menu Item
Added a new menu item that allows content editors to specify an SVG
element that is rendered on the front-end. Mimicing the image menu item,
the user can specify either a content item or an external URL that as
the destination where a user will be taken to when clicking a link.